Python and JS/TS SDKs: Update usage of server_url
First, we can remove the breaking change notices added in
https://github.com/Unstructured-IO/docs/pull/485. This is now considered a bug that was resolved in
`unstructured-client==0.30.2`. Users can pass `server_url` in the client constructor as before.

Next, let's remove the mentions of `server_url` in the platform and serverless SDK examples. We
want to enable the SDK to do the right thing by default, and to provide a unified SaaS
experience across our services. If the user is specifying the platform or serverless URL, they will
override the default behavior, and potentially see errors when they try to mix and match functions.

The `server_url` should be a special case parameter for partitioning users who are not on
Serverless, i.e. free tier or marketplace deploys. We can add a section for these users under
"Customizing the client" in the partitioning section of the docs.